tests: add validation of info requirement endpoints

+++ src/main/java/eu/bbmri_eric/negotiator/service/InformationRequirementServiceImpl.java
@@ -80,6 +80,9 @@ public class InformationRequirementServiceImpl implements InformationRequirement
   @Override
   public void deleteInformationRequirement(Long id) {
     Objects.requireNonNull(id, "InformationRequirement ID must not be null");
+    if (!requirementRepository.existsById(id)) {
+      throw new EntityNotFoundException(id);
+    }
     requirementRepository.deleteById(id);
   }
 }
